can update/delete only own posts

+++ app/routers/post.py
@@ -10,8 +10,9 @@ router = APIRouter(
 )
 
 @router.get("/", response_model=List[schemas.Post])   #order of paths coded matters. will retrieve first matching path oprn
-def get_posts(db: Session = Depends(get_db), user_id: int = Depends(oauth2.get_current_user)):
+def get_posts(db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)): #data type of current_user doesnt matter here
     posts = db.query(models.Post).all()                 #.all sends the query
+    #posts = db.query(models.Post).filter(models.Post.id == current_user.id).all()  #if you want to retrieve only the logged in users own posts
     return posts
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model= schemas.Post)  
@@ -33,8 +34,14 @@ def get_post(id: int, db: Session = Depends(get_db), current_user: int = Depends
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_post(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     post_query = db.query(models.Post).filter(models.Post.id == id)
+    post = post_query.first()
+
     if post_query.first() == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id {id} not found")
+    
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform requested action.")
+
     post_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT) #dont return json for http204 in fastapi
@@ -43,8 +50,13 @@ def delete_post(id: int, db: Session = Depends(get_db), current_user: int = Depe
 def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
+
     if post == None:
         raise HTTPException(status_code= status.HTTP_404_NOT_FOUND, detail=f"post with id {id} not found")
+    
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform requested action.")
+
     post_query.update(updated_post.model_dump(), synchronize_session=False) 
     db.commit()
     return post
